add DELETE & EDIT User Profile functionality
-created update-form component to fill in and pass updated fields except username

+++ src/components/profile-view/profile-view.jsx
@@ -1,30 +1,67 @@
-import React from 'react';
+import React, { useState } from 'react';
 import axios from 'axios';
 import { Button, Container, Row } from 'react-bootstrap';
 import { FavoriteCard } from '../favorite-card/favorite-card';
+import { UpdateForm } from '../update-form/update-form';
+import { connect } from 'react-redux';
 import './profile-view.scss';
 
-export const ProfileView = (props) => {
-  const { user, goBack, movies, handleFavorite } = props;
+// Redux Actions
+import { updateUser, deleteUser } from '../../actions/actions';
+
+function ProfileView(props) {
+  let [showForm, setShowForm] = useState(false);
+  const { handleFavorite, goBack, movies, updateUser, user } = props;
   const { username, email, birthday, favoriteMovies } = user;
-  const token = localStorage.getItem('token');
 
-  const deleteAccount = () => {
-    // Implement deleteUser reducer
-    // ...
+  const handleDeleteUser = () => {
+    const accessToken = localStorage.getItem('token');
+    if (username && accessToken) {
+      let sure = confirm(
+        'Are you sure? This action is irreversible and will ERASE your account.'
+      );
+      if (!sure) return;
+      // request to Delete user from webserver
+      axios
+        .delete(`https://top-flix.herokuapp.com/users/${email}`, {
+          headers: { Authorization: `Bearer ${accessToken}` },
+        })
+        .then((res) => {
+          // Clear Token from local storage
+          localStorage.clear();
+          alert('Your user account was deleted.');
+          deleteUser({});
+          window.open('/', '_self');
+        })
+        .catch((err) => console.log(err));
+    }
+  };
 
-    // Delete user from webserver
+  const handleUpdateUser = (updatedUser) => {
+    console.log('Updated user from handleUpdateUser: ', updatedUser);
+    const { username } = updatedUser;
+    const accessToken = localStorage.getItem('token');
+    // Dispatch updateUser() action
+
+    // Update user data in webserver
     axios
-      .delete(`https://top-flix.herokuapp.com/users/${username}`, {
-        headers: { Authorization: `Bearer ${token}` },
+      .put(
+        `https://top-flix.herokuapp.com/users/${username}`,
+        { ...updatedUser },
+        {
+          Headers: { Authorization: `Bearer ${accessToken}` },
+        }
+      )
+      .then((response) => {
+        const data = response.data;
+        console.log('response data from axios in handleUpdateUser: ', data);
+        document.alert('User info updated');
+        window.open(`/users/${username}`, '_self');
       })
-      .then((res) => {
-        alert(`Your user account was deleted.`);
-        // Clear Token from local storage
-        localStorage.clear();
-        window.open('/', '_self');
-      })
-      .catch((err) => console.log(err));
+      .catch((err) => {
+        console.log('error updating the user:', err);
+      });
+    window.open('/', '_self');
   };
 
   return (
@@ -32,23 +69,47 @@ export const ProfileView = (props) => {
       <h1>
         Profile of <span className="text-info">{username}</span>
       </h1>
-      <Button className="mb-4" variant="warning" onClick={goBack}>
-        « Back
-      </Button>
+      <div className="d-flex">
+        <Button className="mb-4" variant="warning" onClick={goBack}>
+          « Back
+        </Button>
+        <Button
+          className="mb-4 ml-2"
+          variant="info"
+          onClick={() => {
+            setShowForm(!showForm);
+          }}
+        >
+          {!showForm ? 'UPDATE Profile' : 'SHOW Profile'}
+        </Button>
+        <Button
+          className="mb-4 ml-2"
+          variant="danger"
+          onClick={handleDeleteUser}
+        >
+          DELETE Profile
+        </Button>
+      </div>
+      {!showForm ? (
+        <>
+          <h3>
+            Email: <span className="text-info fw-bold ml-4">{email}</span>
+          </h3>
+          <h3>
+            Birthday:{' '}
+            <span className="text-info fw-bold">{`${birthday.slice(
+              5,
+              7
+            )}-${birthday.slice(8, 10)}-${birthday.slice(0, 4)}`}</span>
+            <span id="mini" className="ml-2">
+              (mm-dd-yyyy)
+            </span>
+          </h3>
+        </>
+      ) : (
+        <UpdateForm user={user} handleUpdateUser={handleUpdateUser} />
+      )}
 
-      <h3>
-        Email: <span className="text-info fw-bold ml-4">{email}</span>
-      </h3>
-      <h3>
-        Birthday:{' '}
-        <span className="text-info fw-bold">{`${birthday.slice(
-          5,
-          7
-        )}-${birthday.slice(8, 10)}-${birthday.slice(0, 4)}`}</span>
-        <span id="mini" className="ml-2">
-          (mm-dd-yyyy)
-        </span>
-      </h3>
       <h2 className="subtitle mt-4">LIST OF ♥️ MOVIES:</h2>
       {favoriteMovies.length !== 0 ? (
         <Row className="justify-content-center mt-3">
@@ -74,4 +135,15 @@ export const ProfileView = (props) => {
       )}
     </Container>
   );
+}
+
+const mapStateToProps = (state) => {
+  return {
+    movies: state.movies,
+    user: state.user,
+  };
 };
+
+export default connect(mapStateToProps, { deleteUser, updateUser })(
+  ProfileView
+);
